requirements.txt
pytest~=7.2.2
filelock~=3.12.3
PyYAML~=6.0
pytest-xdist